+++ changelog.md
@@ -1,3 +1,6 @@
 ### 2020-10-26
 - added: BatchNormRelu, Conv2DBatchNorm, ShortCut and Residual Layer
 - added: SimpleResidual file for residual networks
+
+### 2020-12-29
+- added: a new basic block for resnet